P1/app/requirements.txt
fastapi[standard]
uvicorn
PyJWT
requests
python-dotenv